% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacc20812100-package.R
\docType{package}
\name{cacc20812100-package}
\alias{cacc20812100}
\alias{cacc20812100-package}
\title{cacc20812100: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
